MVP complete.
Redis-backed caching implemented.

Images without faces aren't stored or sent to the client. Instead, we
send a plain-text message indicating that no faces were found.
The rationale is threefold:
  * Less data going through our server
  * Prevents someone from abusing us an image cache
  * Minimizes legal attack-surface (we're not storing any unmodified
    images, just ones with faces we've drawn over).
Clients should check the MIME type or length of the response to see if
an actual image was returned, and provide the original image as a
fallback.