URLs really URLs, HTML5 history, works in noscript.
URLs are genuine uniform resource locators, so you can link this to your
friends (long, but all the kids use social networks or URL shorteners
these days, and this approach let me implement it in about 1.5 hours).

Added HTML5 history support, so in more modern browsers you can keep the
URL box (and you're treated to fancy animations) and still have the
browser UI work as expected. Modernizr ensures that we don't attempt
this in browsers that don't support it.

Added a default form action (which is overrided in JS) so it still works
for clients without javascript or those who have it disabled by default.