Redirect to original if no faces, new jade API.
If the server responds with a no faces message, then change the source
attribute of the image to the original URL, rather than re-serving the
image ourselves. This involved fixing a bug in the frontend JS where the
image url would be changed back by swapImage if it fired after the
status check was completed.

Load the views/index.jade template at server startup and compile a
rendering function, in keeping with the Jade 1.5 API.