Cache now Stats, since that's all I'm using it for.
The lib/cache.js redis-backed cache is now lib/stats.js, as stats are
all redis is being used for, now that images are being stored in the
filesystem. All the cache-related operations have been removed, leaving
only the configuration function and Stats.hit, which is called whenever
an image is requested to update the stats related to that image.