Mipmapping implemented.
Since node-canvas's image rescaling isn't the greatest, the faces are
now mipmapped ahead of time using the resources/faces/mipmap.sh script.
This generates scaled versions of the faces, whose heights are powers of
two between 8 and 512, inclusive (the range of powers of 2 can be
changed; see mipmap.sh for details).

The ragefaces.js module has been rewritten, in part due to the
mipmapping. The module's only export is the `face` function. To obtain a
random rageface, just call it without arguments (the returned face will
have a height of 256 pixels). If you know how high you want the face to
be, pass that as the first argument, and the smallest mipmap whose height
is not less than the specified size will be returned (e.g., if you call
`face(37)`, then the returned face will have a height of 64). If you
want a particular face, then pass the name of that face as the second
argument (e.g. `face(143, 'angry')`).