In-flight request tracking.
A global keeps track of in-flight requests. If no cached file is found,
this global is checked to see if there is currently a request for the
desired image in progress. If so, the "wrapped response" (which is just
the response plus whether or not to send an image to the response) is
added to the list at the desired image's key in the global, and then the
function returns. If not, then that list is initialized and the normal
request path continues.

When the normal request path completes, it adds its wrapped response
argument to the list of subscribers for the image that was enraged (the
list in the global mentioned above). It then iterates over the resulting
list and responds to each response object found in it as appropriate.

The net effect of this is to ensure we do minimal work for each cache
miss.