Superifical variable name change.
Instead of using the boolean sendImage key of a wrapped response object
to decide whether to send it an image or text, use the string expects
key of the wrapped respose; if it's 'image', send an image, if it's
'text', send html.

Right now, no big diff, but the latter is more extensible.