New URI paths, fix in-flight request tracking bug.
All image requests should be of the form /image/?src=$img_url. This is
so nginx can reverse proxy image requests and html requests to different
f7u12rl servers, allowing the server that's enraging the images to be
seperated from the server that's actually serving HTML. This offers a
big performance boost, as the site will be able to sustain several
hundred html requests per second, even if the image server's event loop
is being blocked by the enraging operation.

I forgot that push() returns the pushed object, and not the array, so
what was supposed to be the array of responses to complete when an image
was enraged turned out to be just the last one, which meant that none of
the responses were actually completed, and would just timeout. Oops.