Configure nginx to reverse-proxy two f7u12rl procs.
The conf/f7u12rl.nginx.conf defines a server that reverse-proxies to two
different f7u12rl server processes. The frontend process, which listens
on port 63250, just serves the index file (which is rendered using
jade). The backend process handles image and status requests. The
benefit of this seperation is that serving the index never has to wait
on the CPU-intensive, event-loop-blocking step of compositing the rage
faces and the original image.

The frontend server can easily handle several hundred requests per
second, so the site should always be serving up something to the user,
though they might have to wait a second or two for the image if it's
new. The backend can actually do several hundred requests per second as
well if it's mostly serving cached images, but I really have no idea
how many of the requests will be cache misses.