HTTP performance optimizations.
Nginx gzips the html from the frontend server, because why not.

Images are now cacheable, as they have the expires and cache-control
headers set to expire in a day, as well as a last-modified-date which is
set to the last time there was a change in the image drawing in the
enrage.js module (this date must be set manually in enrage.js when it is
updated). Accordingly, sendImg now takes a req argument, checks that
request for an if-modified-since header, and returns a 304 if
appropriate.